+++ packages/core/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @node-escpos/core
 
+## 0.4.0
+
+### Minor Changes
+
+- 7e181a4: Added Character Codepage Table for TM20 and similar models
+- 7d6c597: Added receipt enhancement functions to set top and bottom logo from NV memory
+
+### Patch Changes
+
+- ecef4e7: Fixes italic mode
+
+## 0.3.0
+
+### Minor Changes
+
+- 21a9560: getStatuses does not reject when data is not returned
+- 8f4c52f: OfflineCauseStatus JSON always shows errors
+
+## 0.2.0
+
+### Minor Changes
+
+- 6a7c49a: Enhanced newline with count parameter
+
+### Patch Changes
+
+- 7d18292: Enhanced typings for statuses
+
+## 0.1.0
+
+### Minor Changes
+
+- afd41ef: Added raw buffer possibility for drawLine function
+- 1b1072d: Added function to set left margin using GS command
+- b294648: Added functionality to set charset inside the printer
+
+## 0.0.4
+
+### Patch Changes
+
+- ca1d497: Remove Legacy Mode. to Simple Mode
+- 439ba01: Remove Legacy Mode in Barcode Function Options. to Simple Mode
+
+## 0.0.3
+
+### Patch Changes
+
+- 9a75999: Move "@types/ndarray" & "@types/qr-image" to Dependencies not Dev.
+
 ## 0.0.2
 
 ### Patch Changes