+++ packages/serialport-adapter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @node-escpos/serialport-adapter
 
+## 1.0.0
+
+### Major Changes
+
+- 5c6f405: Changed Serial.list() to be static as it is not bind to the object
+
+## 0.0.3
+
+### Patch Changes
+
+- 3e509a7: 🐛 Fixed an error indicating that the port was open when open the device by serialport.
+
 ## 0.0.2
 
 ### Patch Changes