+++ packages/usb-adapter/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @node-escpos/usb-adapter
 
+## 0.3.1
+
+### Patch Changes
+
+- 7b29a0e: 🐛 solved some bugs.
+
+## 0.3.0
+
+### Minor Changes
+
+- b2fdbeb: 🚀 expose Device types
+
+## 0.2.1
+
+### Patch Changes
+
+- 2aa8939: 🚀
+
+## 0.2.0
+
+### Minor Changes
+
+- d68acb4: expose listener for attach/detach events
+
+## 0.1.0
+
+### Minor Changes
+
+- 8a1d9bd: Added getDeviceBySerial and outsourced isPrinter function for reusability
+
+## 0.0.3
+
+### Patch Changes
+
+- b0bc617: Solved removing all usb detach listeners
+
 ## 0.0.2
 
 ### Patch Changes